Still no pwm on laser pin D10
But some doc upgrades

+++ Arduino/spindle_control.h
@@ -44,15 +44,15 @@ uint8_t spindle_get_state();
 void spindle_sync(uint8_t state, float rpm);
 
 // Sets spindle running state with direction, enable, and spindle PWM.
-void spindle_set_state(uint8_t state, float rpm); 
+void spindle_set_state(uint8_t state, float rpm);
 
 // Sets spindle PWM quickly for stepper ISR. Also called by spindle_set_state().
 // NOTE: Mega2560 PWM register is 16-bit.
-void spindle_set_speed(uint16_t pwm_value);
+void spindle_set_speed(uint8_t pwm_value);
 
 // Computes Mega2560-specific PWM register value for the given RPM for quick updating.
-uint16_t spindle_compute_pwm_value(float rpm);
-  
+uint8_t spindle_compute_pwm_value(float rpm);
+
 // Stop and start spindle routines. Called by all spindle routines and stepper ISR.
 void spindle_stop();
